make sure to call async handler callback
